Move Window creation logic to the constructor
Lazily creating a Window was causing issues when creating resources
(like the Renderer, for example) which depend on a Window being created.

Such resources may be created before calling Window::Show() which cause
a crash.

This commit fixes that by creating the Window in the constructor. Making
sure that once a Window object is created, it's associated system window
will be as well.