Implement static buffers
Added an implementation of using 2 static buffers, one for vertex data
and the second for index data. This way we will not have to create a new
tiny buffer for each mesh.

Worth noting that the implementation is a bit rough around the edges and
very much WIP. Improvements will be made.