GFX: Deprecate the primitive rendering functions
The primitives cannot be transformed or colored because the Mesh is
built in the function instead of user-side. As more and more properties
get added to the Mesh class, those functions' parameter lists will get
out of hand.

A better solution would be to use DrawMesh() and instead of primitive
rendering functions, provide pre-built Meshes that can be copied,
customized and then passed to DrawMesh() for rendering.