Move the Mesh definition into a separate 'Geometry' library.
Both the Scene Loader (in libIO) and libGFX need that information, so
instead of adding a direct dependency between libIO and libGFX, the
'class Mesh' and its supporting structures have been moved to a separate
Geometry library to break the direct dependency between high-level
libraries. GFX::Mesh has been deprecated in favor of Geometry::Mesh,
which should be used from now on.

The geometric definition (vertices and indices) of 'class Mesh' now
resides in libGeometry. A new class, 'class Object,' has been introduced
in libGFX; this class is meant to replace the deprecated 'class Mesh'.
It hosts the Mesh itself along with its properties (Transform and
Material, for now). A new set of functions has been introduced in
GFX::Renderer for rendering instances of 'class Object', and the
DrawMesh() functions have been deprecated in favor of the new
SubmitObject() functions.