Fix Debug builds on windows
Debug and Release binaries cannot be mixed on Windows due to ABI incompatibility,
this commit appends "d" to the Bullet library names on Debug builds so that the project
is correctly linked against the Debug libraries in Debug configurations.

This, of course, assumes that Bullet's Debug libraries have the "d" suffix. This can be done
by using the -DCMAKE_DEBUG_POSTFIX="d" commandline option when running the CMake
configure step when building Bullet.