GFX: Fix command queue overflow
The renderer didn't handle the case where the number of submitted objects
might overflow the static buffers. This commit checks for that and flushes
the render queue when it fills up before submitting new objects.

It also moves the framebuffer clear to the Clear() funnction as clearing it on
each Flush would mean only the last flush's results where actually on screen
since the previous ones would get cleared before the buffers were swapped